Check the size of event content
According to https://github.com/superfeedr/documentation/issues/19,
there might be no Items in a notification and that usually means that
there've been issues in parsing the feed.

By checking the size, and passing null as the Items, the consumer will
have the chance to plan accordingly (e.g., marking, and unsubscribing
from the feed). In its current form it throws a RuntimeException.